Making sure numbers are floats and moving the divide by zero check to be before Radar update
Ex sin(pi)/3 -> sin(pi)/3.0

Checks for dividing by zero right before the Jacobian

+++ src/kalman_filter.cpp
@@ -45,7 +45,7 @@ void KalmanFilter::Update(const VectorXd &z) {
 
   //new estimate
   x_ = x_ + (K * y);
-  long x_size = x_.size();
+  int x_size = x_.size();
   MatrixXd I = MatrixXd::Identity(x_size, x_size);
   P_ = (I - K * H_) * P_;
   
@@ -61,11 +61,12 @@ void KalmanFilter::UpdateEKF(const VectorXd &z) {
   
   VectorXd y = z - z_pred;
   while (y(1) < -pi){
-    y(1) = y(1) + 2*pi;
+    y(1) = y(1) + 2.0*pi;
   }
   while(y(1) > pi){
-    y(1) = y(1) - 2*pi;
+    y(1) = y(1) - 2.0*pi;
   }
+ 
   MatrixXd Ht = H_.transpose();
   MatrixXd S = H_ * P_ * Ht + R_;
   MatrixXd Si = S.inverse();
